+++ stellar_rust_sdk/Cargo.toml
@@ -32,3 +32,4 @@ hex = "0.4.3"
 base64 = "0.21.4"
 chrono = "0.4.31"
 lazy_static = "1.4.0"
+stellar_rust_sdk_derive = { path = "../stellar_rust_sdk_derive" }
