+++ stellar_rust_sdk/src/assets/all_assets_request.rs
