+++ stellar_rust_sdk/src/lib.rs
@@ -502,61 +502,3 @@ pub trait Paginatable {
     where
         Self: Sized;
 }
-
-// #[macro_export]
-// macro_rules! impl_paginatable {
-//     ($struct_name:ident) => {
-//         impl Paginatable for $struct_name {
-//             fn set_cursor(&mut self, cursor: Option<u32>) -> Result<(), String> {
-//                 if cursor == Some(0) {
-//                     Err("Cursor cannot be zero.".to_string())
-//                 } else {
-//                     self.cursor = cursor;
-//                     Ok(())
-//                 }
-//             }
-
-//             fn set_limit(&mut self, limit: Option<u8>) -> Result<(), String> {
-//                 if limit > Some(200) || limit == Some(0) {
-//                     Err("Limit must be between 1 and 200.".to_string())
-//                 } else {
-//                     self.limit = limit;
-//                     Ok(())
-//                 }
-//             }
-
-//             fn set_order(&mut self, order: Option<Order>) -> Result<(), String> {
-//                 // Assuming you always accept the order, but you could add logic here
-//                 self.order = order;
-//                 Ok(())
-//             }
-//         }
-//     };
-// }
-
-#[macro_export]
-macro_rules! impl_paginatable {
-    ($struct_name:ident) => {
-        impl Paginatable for $struct_name {
-            fn set_cursor(self, cursor: u32) -> Result<Self, String> {
-                // Always accept the cursor since it's non-optional in the setter
-                Ok(Self { cursor: Some(cursor), ..self })
-            }
-
-            fn set_limit(self, limit: u8) -> Result<Self, String> {
-                // Validate limit if necessary
-                if !(1..=200).contains(&limit) {
-                    Err("Limit must be between 1 and 200.".to_string())
-                } else {
-                    Ok(Self { limit: Some(limit), ..self })
-                }
-            }
-
-            fn set_order(self, order: Order) -> Result<Self, String> {
-                // No validation required for setting the order in this context
-                Ok(Self { order: Some(order), ..self })
-            }
-        }
-    };
-}
-
