+++ stellar_rust_sdk/src/operations/mod.rs
@@ -1,4 +1,3 @@
-use crate::impl_paginatable;
 
 use self::operations_for_account_request::OperationsForAccountRequest;
