+++ stellar_rust_sdk_derive/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "stellar_rust_sdk_derive"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+quote = "1.0"
+proc-macro2 = "1.0"
+syn = "2.0"
+
+[lib]
+proc-macro = true
